requirements.txt
# AI
google-cloud-aiplatform>=1.45.0
vertexai>=1.0.0

# Browser automation
playwright==1.43.0

# Database
sqlalchemy>=2.0.0
pg8000>=1.30.0
cloud-sql-python-connector[pg8000]>=1.6.0

# Google APIs
google-auth>=2.28.0
google-api-python-client>=2.120.0
google-auth-oauthlib>=1.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
tenacity>=8.2.0
loguru>=0.7.2
httpx>=0.27.0

# Scheduling / async
schedule>=1.2.1
asyncio-mqtt>=0.16.2

# Reporting
jinja2>=3.1.3
weasyprint>=62.0
